+++ libs/core/src/state/mod.ts
@@ -2,7 +2,6 @@ export * from "./theme/mod.ts";
 export * from "./settings/mod.ts";
 export * from "./weg_items.ts";
 export * from "./wm_layout.ts";
-export * from "./placeholder.ts";
 export * from "./icon_pack.ts";
 export * from "./plugin/mod.ts";
 export * from "./widget/mod.ts";