+++ libs/core/src/state/placeholder.rs
@@ -7,7 +7,44 @@ use url::Url;
 
 use crate::{resource::PluginId, utils::TsUnknown};
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema, TS)]
+#[ts(repr(enum = name))]
+pub enum ToolbarJsScope {
+    Date,
+    Notifications,
+    Media,
+    Network,
+    Keyboard,
+    User,
+    Bluetooth,
+    Power,
+    FocusedApp,
+    Workspaces,
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize, JsonSchema, TS)]
+#[serde(default, rename_all = "camelCase")]
+pub struct ToolbarItem {
+    /// Id to identify the item, should be unique. Preferably a uuid.
+    pub id: String,
+    /// List of scopes to be loaded in the item js execution scope.
+    pub scopes: HashSet<ToolbarJsScope>,
+    /// JS function definition for content to display in the item.
+    pub template: String,
+    /// JS function definition for content to display in tooltip of the item.
+    pub tooltip: Option<String>,
+    /// JS function definition badge content, will be displayed over the item, useful as notifications.
+    pub badge: Option<String>,
+    /// JS function definition that will be executed when the item is clicked.
+    #[serde(alias = "onClickV2")]
+    pub on_click: Option<String>,
+    /// Styles to be added to the item. This follow the same interface of React's `style` prop.
+    pub style: HashMap<String, Option<StyleValue>>,
+    /// Remote data to be added to the item scope.
+    pub remote_data: HashMap<String, RemoteDataDeclaration>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(rename_all = "camelCase")]
 pub struct RemoteDataDeclaration {
     url: Url,
@@ -15,65 +52,14 @@ pub struct RemoteDataDeclaration {
     update_interval_seconds: Option<u32>,
 }
 
-macro_rules! common_item {
-    (
-        $(
-            $(#[$scope:meta])*
-            struct $name:ident {
-                $($rest:tt)*
-            }
-        )*
-    ) => {
-        $(
-            #[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
-            #[serde(default, rename_all = "camelCase")]
-            pub struct $name {
-                /// Id to identify the item, should be unique.
-                pub id: String,
-                /// Content to display in the item.
-                ///
-                /// Should follow the [mathjs expression syntax](https://mathjs.org/docs/expressions/syntax.html).
-                ///
-                $(#[$scope])*
-                pub template: String,
-                /// Content to display in tooltip of the item.
-                ///
-                /// Should follow the [mathjs expression syntax](https://mathjs.org/docs/expressions/syntax.html).
-                ///
-                $(#[$scope])*
-                pub tooltip: Option<String>,
-                /// Badge will be displayed over the item, useful as notifications.
-                ///
-                /// Should follow the [mathjs expression syntax](https://mathjs.org/docs/expressions/syntax.html).
-                ///
-                $(#[$scope])*
-                pub badge: Option<String>,
-                /// Deprecated use `onClickV2` instead.
-                pub on_click: Option<String>,
-                /// This code will be parsed and executed when the item is clicked.
-                ///
-                /// Should follow the [mathjs expression syntax](https://mathjs.org/docs/expressions/syntax.html).
-                ///
-                $(#[$scope])*
-                pub on_click_v2: Option<String>,
-                /// Styles to be added to the item. This follow the same interface of React's `style` prop.
-                pub style: HashMap<String, Option<StyleValue>>,
-                /// Remote data to be added to the item scope.
-                pub remote_data: HashMap<String, RemoteDataDeclaration>,
-                $($rest)*
-            }
-        )*
-    };
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(untagged)]
 pub enum StyleValue {
     String(String),
     Number(serde_json::Number),
 }
 
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Clone, Default, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(rename_all = "camelCase")]
 #[ts(repr(enum = name))]
 pub enum WorkspaceToolbarItemMode {
@@ -83,344 +69,27 @@ pub enum WorkspaceToolbarItemMode {
     Numbered,
 }
 
-common_item! {
-    /// ## Base Item Scope
-    /// Have all icons defined on [React Icons](https://react-icons.github.io/react-icons) as properties of the object.
-    /// ```js
-    /// const icon: object;
-    /// ```
-    /// Haves all environment variables defined on the system as properties of the object.
-    /// ```js
-    /// const env: object;
-    /// ```
-    /// Functions to add images to the item.
-    /// ```js
-    /// function getIcon(name: string, size: number = 16): string
-    /// function imgFromUrl (url: string, size: number = 16): string
-    /// function imgFromPath (path: string, size: number = 16): string
-    /// function imgFromExe (exe_path: string, size: number = 16): string
-    /// function t(path: string): string
-    /// ```
-    struct TextToolbarItem {}
-
-    /// ## Generic Item Scope
-    /// ```ts
-    /// // the current focused window
-    /// const window: {
-    ///     name: string;
-    ///     title: string;
-    ///     exe: string | null;
-    /// };
-    /// ```
-    struct GenericToolbarItem {}
-
-    /// ## Date Item Scope
-    /// ```ts
-    /// const date: string; // the formatted date
-    /// ```
-    struct DateToolbarItem {}
-
-    /// ## Power Item Scope
-    /// ```ts
-    /// interface PowerStatus {
-    ///     acLineStatus: number;
-    ///     batteryFlag: number;
-    ///     batteryLifePercent: number;
-    ///     systemStatusFlag: number;
-    ///     batteryLifeTime: number;
-    ///     batteryFullLifeTime: number;
-    /// }
-    ///
-    /// enum PowerPlan {
-    ///   Balanced = 'Balanced',
-    ///   BatterySaver = 'BatterySaver',
-    ///   BetterBattery = 'BetterBattery',
-    ///   GameMode = 'GameMode',
-    ///   HighPerformance = 'HighPerformance',
-    ///   MaxPerformance = 'MaxPerformance',
-    ///   MixedReality = 'MixedReality',
-    /// }
-    ///
-    /// interface Battery {
-    ///     // Static info
-    ///     vendor: string | null;
-    ///     model: string | null;
-    ///     serialNumber: string | null;
-    ///     technology: string;
-    ///
-    ///     // Common information
-    ///     state: string;
-    ///     capacity: number;
-    ///     temperature: number | null;
-    ///     percentage: number;
-    ///     cycleCount: number | null;
-    ///     smartCharging: boolean;
-    ///
-    ///     // Energy stats
-    ///     energy: number;
-    ///     energyFull: number;
-    ///     energyFullDesign: number;
-    ///     energyRate: number;
-    ///     voltage: number;
-    ///
-    ///     // Charge stats
-    ///     timeToFull: number | null;
-    ///     timeToEmpty: number | null;
-    /// }
-    ///
-    /// const power: PowerStatus;
-    /// const powerPlan: PowerPlan;
-    /// const batteries: Battery[];
-    /// const battery: Battery | null;
-    /// ```
-    struct PowerToolbarItem {}
-
-    /// ## Keyboard Item Scope
-    /// ```ts
-    /// interface KeyboardLayout {
-    ///   id: string;
-    ///   handle: string;
-    ///   displayName: string;
-    ///   active: boolean;
-    /// }
-    ///
-    /// interface SystemLanguage {
-    ///   id: string;
-    ///   code: string;
-    ///   name: string;
-    ///   nativeName: string;
-    ///   inputMethods: KeyboardLayout[];
-    /// }
-    ///
-    /// const languages: Language[];
-    /// const activeLang: Language;
-    /// const activeKeyboard: KeyboardLayout;
-    /// const activeLangPrefix: string;
-    /// const activeKeyboardPrefix: string;
-    /// ```
-    struct KeyboardToolbarItem {}
-
-    /// ## Bluetooth Item Scope
-    /// ```ts
-    /// interface BluetoothDevice
-    /// {
-    ///     id: string,
-    ///     name: string,
-    ///     address: bigint,
-    ///     majorClass: BluetoothMajor,
-    ///     minorMainClass: BluetoothMinor,
-    ///     minorSubClass: BluetoothMinor,
-    ///     connected: boolean,
-    ///     paired: boolean,
-    ///     canPair: boolean,
-    ///     isBluetoothLoweenergy: boolean,
-    ///     iconPath: string,
-    /// }
-    ///
-    /// type BluetoothMajor = "Miscellaneous" | "Computer" | "Phone" | "NetworkAccessPoint" | "AudioVideo" | "Peripheral" | "Imaging" | "Wearable" | "Toy" | "Health" | "Unkown";
-    ///
-    /// type BluetoothMinor = "Uncategorized" | "ComputerDesktop" | "ComputerServer" | "ComputerLaptop" | "ComputerHandheld" | "ComputerPalmSize" | "ComputerWearable" | "ComputerTablet"
-    ///     | "PhoneCellular" | "PhoneCordless" | "PhoneSmartPhone" | "PhoneWired" | "PhoneIsdn" | "NetworkFullyAvailable" | "NetworkUsed01To17Percent" | "NetworkUsed17To33Percent" | "NetworkUsed33To50Percent"
-    ///     | "NetworkUsed50To67Percent" | "NetworkUsed67To83Percent" | "NetworkUsed83To99Percent" | "NetworkNoServiceAvailable" | "AudioVideoWearableHeadset" | "AudioVideoHandsFree" | "AudioVideoMicrophone"
-    ///     | "AudioVideoLoudspeaker" | "AudioVideoHeadphones" | "AudioVideoPortableAudio" | "AudioVideoCarAudio" | "AudioVideoSetTopBox" | "AudioVideoHifiAudioDevice" | "AudioVideoVcr" | "AudioVideoVideoCamera"
-    ///     | "AudioVideoCamcorder" | "AudioVideoVideoMonitor" | "AudioVideoVideoDisplayAndLoudspeaker" | "AudioVideoVideoConferencing" | "AudioVideoGamingOrToy" | "PeripheralJoystick" | "PeripheralGamepad"
-    ///     | "PeripheralRemoteControl" | "PeripheralSensing" | "PeripheralDigitizerTablet" | "PeripheralCardReader" | "PeripheralDigitalPen" | "PeripheralHandheldScanner" | "PeripheralHandheldGesture"
-    ///     | "WearableWristwatch" | "WearablePager" | "WearableJacket" | "WearableHelmet" | "WearableGlasses" | "ToyRobot" | "ToyVehicle" | "ToyDoll" | "ToyController" | "ToyGame" | "HealthBloodPressureMonitor"
-    ///     | "HealthThermometer" | "HealthWeighingScale" | "HealthGlucoseMeter" | "HealthPulseOximeter" | "HealthHeartRateMonitor" | "HealthHealthDataDisplay" | "HealthStepCounter" | "HealthBodyCompositionAnalyzer"
-    ///     | "HealthPeakFlowMonitor" | "HealthMedicationMonitor" | "HealthKneeProsthesis" | "HealthAnkleProsthesis" | "HealthGenericHealthManager" | "HealthPersonalMobilityDevice" | "PeripheralOther" | "PeripheralPointer"
-    ///     | "PeripheralKeyboard" | "PeripheralKeyboardAndPointer";
-    ///
-    /// const bluetoothState: boolean;
-    /// const devices: BluetoothDevice[];
-    /// const connectedDevices: BluetoothDevice[];
-    /// ```
-    struct BluetoothToolbarItem {
-        /// Show bluetooth selector popup on click]
-        #[serde(default)]
-        with_bluetooth_selector: bool,
-    }
-
-    /// ## Network Item Scope
-    /// ```ts
-    /// interface NetworkInterface {
-    ///     name: string;
-    ///     description: string;
-    ///     status: 'up' | 'down';
-    ///     dnsSuffix: string;
-    ///     type: string;
-    ///     gateway: string | null;
-    ///     mac: string;
-    ///     ipv4: string | null;
-    ///     ipv6: string | null;
-    /// }
-    /// const online: boolean;
-    /// const interfaces: NetworkInterface[];
-    /// const usingInterface: NetworkInterface | null;
-    /// ```
-    struct NetworkToolbarItem {
-        /// Show Wi-fi selector popup on click]
-        #[serde(default)]
-        with_wlan_selector: bool,
-    }
-
-    /// ## Media Item Scope
-    /// ```ts
-    /// const volume: number; // output master volume from 0 to 1
-    /// const isMuted: boolean; // output master volume is muted
-    /// const inputVolume: number; // input master volume from 0 to 1
-    /// const inputIsMuted: boolean; // input master volume is muted
-    ///
-    /// interface MediaSession {
-    ///     id: string;
-    ///     title: string;
-    ///     author: string;
-    ///     thumbnail: string | null; // path to temporal media session image
-    ///     playing: boolean;
-    ///     default: boolean;
-    ///     owner: {
-    ///         name: string;
-    ///         iconPath: string | null;
-    ///     } | null;
-    /// }
-    ///
-    /// const mediaSession: MediaSession | null;
-    /// ```
-    struct MediaToolbarItem {
-        /// Show media controls popup on click
-        #[serde(default)]
-        with_media_controls: bool,
-    }
-
-    /// ## Notifications Item Scope
-    /// ```ts
-    /// const count: number;
-    /// ```
-    struct UserToolbarItem {
-        /// Show user control popup on click
-        #[serde(default)]
-        with_user_folder: bool,
-    }
-
-
-    /// ## Notifications Item Scope
-    /// ```ts
-    /// const count: number;
-    /// ```
-    struct NotificationsToolbarItem {}
-
-    /// ## Device Item Scope
-    /// this module does no expand the scope of the item
-    struct DeviceToolbarItem {}
-
-    /// ## Settings Item Scope
-    /// this module does no expand the scope of the item
-    struct SettingsToolbarItem {}
-
-    /// ## Workspace Item Scope
-    /// this module does no expand the scope of the item
-    struct WorkspaceToolbarItem {
-        mode: WorkspaceToolbarItemMode,
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
-#[serde(tag = "type")]
-pub enum ToolbarItem {
-    #[serde(alias = "text")]
-    Text(TextToolbarItem),
-    #[serde(alias = "generic")]
-    Generic(GenericToolbarItem),
-    #[serde(alias = "date")]
-    Date(DateToolbarItem),
-    #[serde(alias = "power")]
-    Power(PowerToolbarItem),
-    #[serde(alias = "keyboard")]
-    Keyboard(KeyboardToolbarItem),
-    #[serde(alias = "network")]
-    Network(NetworkToolbarItem),
-    #[serde(alias = "bluetooth")]
-    Bluetooth(BluetoothToolbarItem),
-    #[serde(alias = "media")]
-    Media(MediaToolbarItem),
-    #[serde(alias = "user")]
-    User(UserToolbarItem),
-    #[serde(alias = "notifications")]
-    Notifications(NotificationsToolbarItem),
-    #[serde(alias = "device")]
-    Device(DeviceToolbarItem),
-    #[serde(alias = "settings")]
-    Settings(SettingsToolbarItem),
-    #[serde(alias = "workspaces")]
-    Workspaces(WorkspaceToolbarItem),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
-#[cfg_attr(feature = "gen-binds", ts(export, repr(enum = name)))]
-pub enum ToolbarModuleType {
-    Text,
-    Generic,
-    Date,
-    Power,
-    Keyboard,
-    Network,
-    Bluetooth,
-    Media,
-    User,
-    Notifications,
-    Device,
-    Settings,
-    Workspaces,
-}
-
 impl ToolbarItem {
     pub fn id(&self) -> String {
-        match self {
-            ToolbarItem::Text(item) => item.id.clone(),
-            ToolbarItem::Generic(item) => item.id.clone(),
-            ToolbarItem::Date(item) => item.id.clone(),
-            ToolbarItem::Power(item) => item.id.clone(),
-            ToolbarItem::Keyboard(item) => item.id.clone(),
-            ToolbarItem::Network(item) => item.id.clone(),
-            ToolbarItem::Bluetooth(item) => item.id.clone(),
-            ToolbarItem::Media(item) => item.id.clone(),
-            ToolbarItem::User(item) => item.id.clone(),
-            ToolbarItem::Notifications(item) => item.id.clone(),
-            ToolbarItem::Device(item) => item.id.clone(),
-            ToolbarItem::Settings(item) => item.id.clone(),
-            ToolbarItem::Workspaces(item) => item.id.clone(),
-        }
+        self.id.clone()
     }
 
     pub fn set_id(&mut self, id: String) {
-        match self {
-            ToolbarItem::Text(item) => item.id = id,
-            ToolbarItem::Generic(item) => item.id = id,
-            ToolbarItem::Date(item) => item.id = id,
-            ToolbarItem::Power(item) => item.id = id,
-            ToolbarItem::Keyboard(item) => item.id = id,
-            ToolbarItem::Network(item) => item.id = id,
-            ToolbarItem::Bluetooth(item) => item.id = id,
-            ToolbarItem::Media(item) => item.id = id,
-            ToolbarItem::User(item) => item.id = id,
-            ToolbarItem::Notifications(item) => item.id = id,
-            ToolbarItem::Device(item) => item.id = id,
-            ToolbarItem::Settings(item) => item.id = id,
-            ToolbarItem::Workspaces(item) => item.id = id,
-        }
+        self.id = id;
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(untagged)]
 pub enum ToolbarItem2 {
     Plugin(PluginId),
     Inline(Box<ToolbarItem>),
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize, Deserialize, JsonSchema, TS)]
+#[derive(Debug, Clone, Default, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(default, rename_all = "camelCase")]
 #[cfg_attr(feature = "gen-binds", ts(export))]
-pub struct Placeholder {
+pub struct ToolbarState {
     /// Whether the reordering possible on the toolbar
     pub is_reorder_disabled: bool,
     /// Items to be displayed in the toolbar
@@ -431,7 +100,7 @@ pub struct Placeholder {
     pub right: Vec<ToolbarItem2>,
 }
 
-impl Placeholder {
+impl ToolbarState {
     fn migrate_plugin_id(id: PluginId) -> PluginId {
         match id.as_str() {
             "@default/system-tray" => "@seelen/tb-system-tray".into(),
@@ -461,6 +130,12 @@ impl Placeholder {
                     }
                 }
                 ToolbarItem2::Inline(mut item) => {
+                    // migration step for old default separator before v2.5
+                    if item.template.contains("window") && item.scopes.is_empty() {
+                        item.scopes.insert(ToolbarJsScope::FocusedApp);
+                        item.template = item.template.replace("window", "focusedApp");
+                    }
+
                     if item.id().is_empty() {
                         item.set_id(uuid::Uuid::new_v4().to_string());
                     }