+++ src/ui/react/toolbar/modules/item/infra/Inner.tsx
@@ -3,31 +3,23 @@ import { CSS } from "@dnd-kit/utilities";
 import { useComputed } from "@preact/signals";
 import type { ToolbarItem } from "@seelen-ui/lib/types";
 import { cx } from "libs/ui/react/utils/styling.ts";
-import { Tooltip } from "antd";
 import type { HTMLAttributes } from "preact/compat";
 
 import { EvaluateAction } from "../app/actionEvaluator.ts";
 import { useItemScope, useRemoteData } from "../app/hooks/index.ts";
 import { $toolbar_state } from "../../shared/state/items.ts";
-import { SanboxedComponent } from "./EvaluatedComponents.tsx";
 import { useItemContextMenu } from "./ContextMenu.tsx";
+import { ElementsFromEvaluated, StringFromEvaluated, useSandboxedCode } from "./EvaluatedComponents.tsx";
 
 export interface InnerItemProps extends HTMLAttributes<HTMLDivElement> {
   module: Omit<ToolbarItem, "type">;
   extraVars?: Record<string, any>;
-  onClick?: (e: MouseEvent) => void;
 }
 
 export function InnerItem(props: InnerItemProps) {
-  const {
-    extraVars = {},
-    module,
-    onClick: onClickProp,
-    children,
-    ...rest
-  } = props;
+  const { extraVars = {}, module, ...rest } = props;
 
-  const { template, tooltip, onClickV2, style, id, badge, remoteData = {} } = module;
+  const { onClick, style, id, remoteData = {} } = module;
 
   const { onContextMenu } = useItemContextMenu(id);
 
@@ -46,48 +38,49 @@ export function InnerItem(props: InnerItemProps) {
     fetchedData,
   });
 
+  const content = useSandboxedCode({ code: module.template, scope });
+  const tooltip = module.tooltip ? useSandboxedCode({ code: module.tooltip, scope }) : null;
+  const badge = module.badge ? useSandboxedCode({ code: module.badge, scope }) : null;
+
+  if (!content) {
+    return null;
+  }
+
   return (
-    <Tooltip
-      arrow={false}
-      mouseLeaveDelay={0}
-      classNames={{ root: "ft-bar-item-tooltip" }}
-      title={tooltip ? <SanboxedComponent code={tooltip} scope={scope} /> : undefined}
+    <div
+      {...rest}
+      id={id}
+      ref={setNodeRef}
+      {...listeners}
+      {...(attributes as HTMLAttributes<HTMLDivElement>)}
+      title={tooltip ? StringFromEvaluated({ content: tooltip }) : undefined}
+      style={{
+        ...style,
+        transform: CSS.Translate.toString(transform),
+        transition,
+        opacity: isDragging ? 0.3 : 1,
+      }}
+      className={cx("ft-bar-item", {
+        "ft-bar-item-clickable": onClick,
+      })}
+      onClick={() => {
+        if (onClick) {
+          EvaluateAction(onClick, scope);
+        }
+      }}
+      onContextMenu={(e: MouseEvent) => {
+        e.stopPropagation();
+        onContextMenu();
+      }}
     >
-      <div
-        {...rest}
-        id={id}
-        ref={setNodeRef}
-        {...listeners}
-        {...(attributes as HTMLAttributes<HTMLDivElement>)}
-        style={{
-          ...style,
-          transform: CSS.Translate.toString(transform),
-          transition,
-          opacity: isDragging ? 0.3 : 1,
-        }}
-        className={cx("ft-bar-item", {
-          "ft-bar-item-clickable": onClickProp || onClickV2,
-        })}
-        onClick={(e: MouseEvent) => {
-          onClickProp?.(e);
-          if (onClickV2) {
-            EvaluateAction(onClickV2, scope);
-          }
-        }}
-        onContextMenu={(e: MouseEvent) => {
-          e.stopPropagation();
-          onContextMenu();
-        }}
-      >
-        <div className="ft-bar-item-content">
-          {children || <SanboxedComponent code={template} scope={scope} />}
-          {!!badge && (
-            <div className="ft-bar-item-badge">
-              <SanboxedComponent code={badge} scope={scope} />
-            </div>
-          )}
-        </div>
+      <div className="ft-bar-item-content">
+        <ElementsFromEvaluated content={content} />
+        {!!badge && (
+          <div className="ft-bar-item-badge">
+            <ElementsFromEvaluated content={badge} />
+          </div>
+        )}
       </div>
-    </Tooltip>
+    </div>
   );
 }